Updated template and script

+++ addDayFromTemplate.sh
@@ -5,9 +5,10 @@ YEAR=$1
 DAY=$(printf "%02d" "$2")
 
 RUST_SRC_FOLDER="./src/year$YEAR/day$DAY"
-FILE_PART1="$RUST_SRC_FOLDER/bin/part1.rs"
-FILE_PART2="$RUST_SRC_FOLDER/bin/part2.rs"
+FILE_PART1="$RUST_SRC_FOLDER/lib/part1.rs"
+FILE_PART2="$RUST_SRC_FOLDER/lib/part2.rs"
 FILE_PARSER="$RUST_SRC_FOLDER/lib/parser.rs"
+FILE_MOD="$RUST_SRC_FOLDER/lib/mod.rs"
 INPUTS_FOLDER_PATH="./input/$YEAR/$DAY/"
 
 if [ -f "$FILE_PART1" ]; then
@@ -25,22 +26,17 @@ if [ -f "$FILE_PARSER" ]; then
     exit 1
 fi
 
-mkdir -p "$RUST_SRC_FOLDER/bin"
+if [ -f "$FILE_MOD" ]; then
+    echo "$FILE_MOD already exists."
+    exit 1
+fi
+
 mkdir -p "$RUST_SRC_FOLDER/lib"
 cp ./template/part.rs.template "$FILE_PART1"
 cp ./template/part.rs.template "$FILE_PART2"
 cp ./template/parser.rs.template "$FILE_PARSER"
+cp ./template/mod.rs.template "$FILE_MOD"
 echo "pub mod day$DAY { pub mod lib; }" >> "src/year$YEAR.rs"
-echo "pub mod parser;" > "$RUST_SRC_FOLDER/lib/mod.rs"
-echo "
-[[bin]]
-name = \"$YEAR-day$DAY-part1\"
-path = \"src/year$YEAR/day$DAY/bin/part1.rs\"
-
-[[bin]]
-name = \"$YEAR-day$DAY-part2\"
-path = \"src/year$YEAR/day$DAY/bin/part2.rs\"" >> Cargo.toml
-
 
 gsed -i "s/<DAY>/$DAY/g" "$FILE_PART1"
 gsed -i "s/<YEAR>/$YEAR/g" "$FILE_PART1"
@@ -48,6 +44,12 @@ gsed -i "s/<PART_NUMBER>/1/g" "$FILE_PART1"
 gsed -i "s/<DAY>/$DAY/g" "$FILE_PART2"
 gsed -i "s/<YEAR>/$YEAR/g" "$FILE_PART2"
 gsed -i "s/<PART_NUMBER>/2/g" "$FILE_PART2"
+gsed -i "s/<DAY>/$DAY/g" "$FILE_PARSER"
+gsed -i "s/<YEAR>/$YEAR/g" "$FILE_PARSER"
+gsed -i "s/<PART_NUMBER>/2/g" "$FILE_PARSER"
+gsed -i "s/<DAY>/$DAY/g" "$FILE_MOD"
+gsed -i "s/<YEAR>/$YEAR/g" "$FILE_MOD"
+gsed -i "s/<PART_NUMBER>/2/g" "$FILE_MOD"
 
 mkdir -p "$INPUTS_FOLDER_PATH"
 touch "$INPUTS_FOLDER_PATH/inputs.txt"
